backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
numpy==1.26.4
pydantic==2.10.0
opencv-python==4.10.0.84
# TensorFlow with GPU support for RTX GPUs (CUDA 11.8+)
tensorflow[and-cuda]==2.15.0
mediapipe==0.10.22
colorthief==0.2.1
